refactor: change filter, sort comonents
