+++ src/utils/baseColumns.ts
@@ -1,0 +1,8 @@
+const baseColumns = [
+  { key: 'year', label: 'Year' },
+  { key: 'population', label: 'Population' },
+  { key: 'co2', label: 'CO2' },
+  { key: 'co2_per_capita', label: 'CO2 per Capita' },
+];
+
+export default baseColumns;
